Performance enhancement to make video thumbnail
To make video thumbnail, it is used to resize and rotate image API.
The image resize/rotate API use gstreamer opensource and it is needed to initialize gstreamer with gst_init().
But it is not fast so that it is used to make video thumbnail.
So resizing/rotation is replaced to new way.
 - resize: mm_util(gstreamer) -> EFL(evas)
 - rotation: mm_util(gstreamer) -> Make new rotate function

Change-Id: I11946578fa7c4087d60d49f7d5afe58fc1a7838c
Signed-off-by: Jiyong Min <jiyong.min@samsung.com>